+++ src/Libraries/SmartStore.Services/Messages/IMessageTokenProvider.cs
@@ -3,6 +3,7 @@ using SmartStore.Core.Domain.Blogs;
 using SmartStore.Core.Domain.Catalog;
 using SmartStore.Core.Domain.Customers;
 using SmartStore.Core.Domain.Forums;
+using SmartStore.Core.Domain.Localization;
 using SmartStore.Core.Domain.Messages;
 using SmartStore.Core.Domain.News;
 using SmartStore.Core.Domain.Orders;
@@ -11,7 +12,7 @@ using SmartStore.Core.Domain.Stores;
 
 namespace SmartStore.Services.Messages
 {
-    public partial interface IMessageTokenProvider
+	public partial interface IMessageTokenProvider
     {
 		void AddStoreTokens(IList<Token> tokens, Store store);
 
@@ -37,7 +38,7 @@ namespace SmartStore.Services.Messages
 
         void AddNewsCommentTokens(IList<Token> tokens, NewsComment newsComment);
 
-		void AddProductTokens(IList<Token> tokens, Product product, int languageId);
+		void AddProductTokens(IList<Token> tokens, Product product, Language language);
 
 		void AddForumTokens(IList<Token> tokens, Forum forum, int languageId);