+++ src/Libraries/SmartStore.Services/Messages/WorkflowMessageService.cs
@@ -8,6 +8,7 @@ using SmartStore.Core.Domain.Catalog;
 using SmartStore.Core.Domain.Common;
 using SmartStore.Core.Domain.Customers;
 using SmartStore.Core.Domain.Forums;
+using SmartStore.Core.Domain.Localization;
 using SmartStore.Core.Domain.Messages;
 using SmartStore.Core.Domain.News;
 using SmartStore.Core.Domain.Orders;
@@ -232,7 +233,7 @@ namespace SmartStore.Services.Messages
             return name ?? customer.Username.EmptyNull();
         }
 
-		protected int EnsureLanguageIsActive(int languageId, int storeId)
+		protected Language EnsureLanguageIsActive(int languageId, int storeId)
         {
 			//load language by specified ID
             var language = _languageService.GetLanguageById(languageId);
@@ -250,7 +251,8 @@ namespace SmartStore.Services.Messages
 
 			if (language == null)
 				throw new Exception("No active language could be loaded");
-            return language.Id;
+
+            return language;
         }
 
         #endregion
@@ -271,7 +273,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("customer");
 
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("NewCustomer.Notification", store.Id);
             if (messageTemplate == null)
@@ -285,17 +287,14 @@ namespace SmartStore.Services.Messages
             //event notification
 			_eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
 
             // use customer email as reply address
 			var replyTo = GetReplyToEmail(customer);
 
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName,
-				replyTo.Item1, replyTo.Item2);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName, replyTo.Item1, replyTo.Item2);
         }
 
         /// <summary>
@@ -310,7 +309,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("customer");
 
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("Customer.WelcomeMessage", store.Id);
             if (messageTemplate == null)
@@ -324,12 +323,11 @@ namespace SmartStore.Services.Messages
             //event notification
 			_eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = customer.Email;
             var toName = customer.GetFullName();
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -344,7 +342,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("customer");
 
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("Customer.EmailValidationMessage", store.Id);
 			if (messageTemplate == null)
@@ -358,12 +356,11 @@ namespace SmartStore.Services.Messages
             //event notification
 			_eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = customer.Email;
             var toName = customer.GetFullName();
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -378,7 +375,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("customer");
 
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("Customer.PasswordRecovery", store.Id);
 			if (messageTemplate == null)
@@ -392,12 +389,11 @@ namespace SmartStore.Services.Messages
             //event notification
 			_eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = customer.Email;
             var toName = customer.GetFullName();
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         #endregion
@@ -416,7 +412,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("order");
 
 			var store = _storeService.GetStoreById(order.StoreId) ?? _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("OrderPlaced.StoreOwnerNotification", store.Id);
 			if (messageTemplate == null)
@@ -425,13 +421,13 @@ namespace SmartStore.Services.Messages
 			//tokens
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
-			_messageTokenProvider.AddOrderTokens(tokens, order, languageId);
+			_messageTokenProvider.AddOrderTokens(tokens, order, language.Id);
 			_messageTokenProvider.AddCustomerTokens(tokens, order.Customer);
 
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
 
@@ -443,10 +439,7 @@ namespace SmartStore.Services.Messages
 				replyToName += ", " + order.BillingAddress.Company;
             }
 
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName,
-				replyToEmail, replyToName);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName, replyToEmail, replyToName);
         }
 
         /// <summary>
@@ -461,7 +454,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("order");
 
 			var store = _storeService.GetStoreById(order.StoreId) ?? _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("OrderPlaced.CustomerNotification", store.Id);
 			if (messageTemplate == null)
@@ -470,7 +463,7 @@ namespace SmartStore.Services.Messages
 			//tokens
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
-			_messageTokenProvider.AddOrderTokens(tokens, order, languageId);
+			_messageTokenProvider.AddOrderTokens(tokens, order, language.Id);
 			_messageTokenProvider.AddCustomerTokens(tokens, order.Customer);
 
             _messageTokenProvider.AddCompanyTokens(tokens);
@@ -480,12 +473,11 @@ namespace SmartStore.Services.Messages
             // event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = order.BillingAddress.Email;
             var toName = string.Format("{0} {1}", order.BillingAddress.FirstName, order.BillingAddress.LastName);
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -504,7 +496,7 @@ namespace SmartStore.Services.Messages
                 throw new Exception("Order cannot be loaded");
 
 			var store = _storeService.GetStoreById(order.StoreId) ?? _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("ShipmentSent.CustomerNotification", store.Id);
             if (messageTemplate == null)
@@ -513,19 +505,18 @@ namespace SmartStore.Services.Messages
 			//tokens
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
-			_messageTokenProvider.AddShipmentTokens(tokens, shipment, languageId);
-			_messageTokenProvider.AddOrderTokens(tokens, shipment.Order, languageId);
+			_messageTokenProvider.AddShipmentTokens(tokens, shipment, language.Id);
+			_messageTokenProvider.AddOrderTokens(tokens, shipment.Order, language.Id);
 			_messageTokenProvider.AddCustomerTokens(tokens, shipment.Order.Customer);
 
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = order.BillingAddress.Email;
             var toName = string.Format("{0} {1}", order.BillingAddress.FirstName, order.BillingAddress.LastName);
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -544,7 +535,7 @@ namespace SmartStore.Services.Messages
                 throw new Exception("Order cannot be loaded");
 
 			var store = _storeService.GetStoreById(order.StoreId) ?? _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("ShipmentDelivered.CustomerNotification", store.Id);
             if (messageTemplate == null)
@@ -553,19 +544,18 @@ namespace SmartStore.Services.Messages
 			//tokens
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
-			_messageTokenProvider.AddShipmentTokens(tokens, shipment, languageId);
-			_messageTokenProvider.AddOrderTokens(tokens, shipment.Order, languageId);
+			_messageTokenProvider.AddShipmentTokens(tokens, shipment, language.Id);
+			_messageTokenProvider.AddOrderTokens(tokens, shipment.Order, language.Id);
 			_messageTokenProvider.AddCustomerTokens(tokens, shipment.Order.Customer);
 
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = order.BillingAddress.Email;
             var toName = string.Format("{0} {1}", order.BillingAddress.FirstName, order.BillingAddress.LastName);
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -580,7 +570,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("order");
 
 			var store = _storeService.GetStoreById(order.StoreId) ?? _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("OrderCompleted.CustomerNotification", store.Id);
 			if (messageTemplate == null)
@@ -589,7 +579,7 @@ namespace SmartStore.Services.Messages
 			//tokens
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
-			_messageTokenProvider.AddOrderTokens(tokens, order, languageId);
+			_messageTokenProvider.AddOrderTokens(tokens, order, language.Id);
 			_messageTokenProvider.AddCustomerTokens(tokens, order.Customer);
 
             _messageTokenProvider.AddCompanyTokens(tokens);
@@ -599,12 +589,11 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = order.BillingAddress.Email;
             var toName = string.Format("{0} {1}", order.BillingAddress.FirstName, order.BillingAddress.LastName);
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -619,7 +608,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("order");
 
 			var store = _storeService.GetStoreById(order.StoreId) ?? _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("OrderCancelled.CustomerNotification", store.Id);
 			if (messageTemplate == null)
@@ -628,18 +617,17 @@ namespace SmartStore.Services.Messages
 			//tokens
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
-			_messageTokenProvider.AddOrderTokens(tokens, order, languageId);
+			_messageTokenProvider.AddOrderTokens(tokens, order, language.Id);
 			_messageTokenProvider.AddCustomerTokens(tokens, order.Customer);
 
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = order.BillingAddress.Email;
             var toName = string.Format("{0} {1}", order.BillingAddress.FirstName, order.BillingAddress.LastName);
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -656,7 +644,7 @@ namespace SmartStore.Services.Messages
             var order = orderNote.Order;
 
 			var store = _storeService.GetStoreById(order.StoreId) ?? _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("Customer.NewOrderNote", store.Id);
 			if (messageTemplate == null)
@@ -666,18 +654,17 @@ namespace SmartStore.Services.Messages
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
 			_messageTokenProvider.AddOrderNoteTokens(tokens, orderNote);
-			_messageTokenProvider.AddOrderTokens(tokens, orderNote.Order, languageId);
+			_messageTokenProvider.AddOrderTokens(tokens, orderNote.Order, language.Id);
 			_messageTokenProvider.AddCustomerTokens(tokens, orderNote.Order.Customer);
 
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = order.BillingAddress.Email;
             var toName = string.Format("{0} {1}", order.BillingAddress.FirstName, order.BillingAddress.LastName);
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -692,7 +679,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("recurringPayment");
 
 			var store = _storeService.GetStoreById(recurringPayment.InitialOrder.StoreId) ?? _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("RecurringPaymentCancelled.StoreOwnerNotification", store.Id);
 			if (messageTemplate == null)
@@ -701,20 +688,18 @@ namespace SmartStore.Services.Messages
 			//tokens
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
-			_messageTokenProvider.AddOrderTokens(tokens, recurringPayment.InitialOrder, languageId);
+			_messageTokenProvider.AddOrderTokens(tokens, recurringPayment.InitialOrder, language.Id);
 			_messageTokenProvider.AddCustomerTokens(tokens, recurringPayment.InitialOrder.Customer);
 			_messageTokenProvider.AddRecurringPaymentTokens(tokens, recurringPayment);
             
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
 
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         #endregion
@@ -727,14 +712,13 @@ namespace SmartStore.Services.Messages
         /// <param name="subscription">Newsletter subscription</param>
         /// <param name="languageId">Language identifier</param>
         /// <returns>Queued email identifier</returns>
-        public virtual int SendNewsLetterSubscriptionActivationMessage(NewsLetterSubscription subscription,
-            int languageId)
+        public virtual int SendNewsLetterSubscriptionActivationMessage(NewsLetterSubscription subscription, int languageId)
         {
             if (subscription == null)
                 throw new ArgumentNullException("subscription");
 
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("NewsLetterSubscription.ActivationMessage", store.Id);
 			if (messageTemplate == null)
@@ -748,12 +732,11 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = subscription.Email;
             var toName = "";
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -762,14 +745,13 @@ namespace SmartStore.Services.Messages
         /// <param name="subscription">Newsletter subscription</param>
         /// <param name="languageId">Language identifier</param>
         /// <returns>Queued email identifier</returns>
-        public virtual int SendNewsLetterSubscriptionDeactivationMessage(NewsLetterSubscription subscription,
-            int languageId)
+        public virtual int SendNewsLetterSubscriptionDeactivationMessage(NewsLetterSubscription subscription, int languageId)
         {
             if (subscription == null)
                 throw new ArgumentNullException("subscription");
 
 			var store = _storeContext.CurrentStore;
-			languageId = EnsureLanguageIsActive(languageId, store.Id);
+			var language = EnsureLanguageIsActive(languageId, store.Id);
 			
             var messageTemplate = GetActiveMessageTemplate("NewsLetterSubscription.DeactivationMessage", store.Id);
             if (messageTemplate == null)
@@ -782,12 +764,11 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = subscription.Email;
             var toName = "";
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         #endregion
@@ -814,7 +795,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("product");
 
 			var store = _storeContext.CurrentStore;
-			languageId = EnsureLanguageIsActive(languageId, store.Id);
+			var language = EnsureLanguageIsActive(languageId, store.Id);
             
 			var messageTemplate = GetActiveMessageTemplate("Service.EmailAFriend", store.Id);
 			if (messageTemplate == null)
@@ -824,19 +805,19 @@ namespace SmartStore.Services.Messages
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
 			_messageTokenProvider.AddCustomerTokens(tokens, customer);
-			_messageTokenProvider.AddProductTokens(tokens, product, languageId);
+			_messageTokenProvider.AddProductTokens(tokens, product, language);
+
 			tokens.Add(new Token("EmailAFriend.PersonalMessage", personalMessage, true));
 			tokens.Add(new Token("EmailAFriend.Email", customerEmail));
 
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = friendsEmail;
             var toName = "";
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         public virtual int SendProductQuestionMessage(Customer customer, int languageId, Product product, 
@@ -852,7 +833,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("product");
 			
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 			
             var messageTemplate = GetActiveMessageTemplate("Product.AskQuestion", store.Id);
             if (messageTemplate == null)
@@ -861,7 +842,7 @@ namespace SmartStore.Services.Messages
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
 			_messageTokenProvider.AddCustomerTokens(tokens, customer);
-			_messageTokenProvider.AddProductTokens(tokens, product, languageId);
+			_messageTokenProvider.AddProductTokens(tokens, product, language);
 
 			tokens.Add(new Token("ProductQuestion.Message", question, true));
             tokens.Add(new Token("ProductQuestion.SenderEmail", senderEmail));
@@ -871,11 +852,11 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
 
-            return SendNotification(messageTemplate, emailAccount, languageId, tokens, toEmail, toName, senderEmail, senderName);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName, senderEmail, senderName);
         }
 
         /// <summary>
@@ -894,7 +875,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("customer");
 
             var store = _storeContext.CurrentStore;
-			languageId = EnsureLanguageIsActive(languageId, store.Id);
+			var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("Wishlist.EmailAFriend", store.Id);
 			if (messageTemplate == null)
@@ -910,12 +891,11 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = friendsEmail;
             var toName = "";
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         #endregion
@@ -935,7 +915,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("returnRequest");
 
 			var store = _storeService.GetStoreById(orderItem.Order.StoreId) ?? _storeContext.CurrentStore;
-			languageId = EnsureLanguageIsActive(languageId, store.Id);
+			var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("NewReturnRequest.StoreOwnerNotification", store.Id);
 			if (messageTemplate == null)
@@ -950,17 +930,14 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
 
             // use customer email as reply address
 			var replyTo = GetReplyToEmail(returnRequest.Customer);
 
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName,
-				replyTo.Item1, replyTo.Item2);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName, replyTo.Item1, replyTo.Item2);
         }
 
         /// <summary>
@@ -976,7 +953,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("returnRequest");
 
 			var store = _storeService.GetStoreById(orderItem.Order.StoreId) ?? _storeContext.CurrentStore;
-			languageId = EnsureLanguageIsActive(languageId, store.Id);
+			var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("ReturnRequestStatusChanged.CustomerNotification", store.Id);
 			if (messageTemplate == null)
@@ -991,14 +968,14 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = returnRequest.Customer.FindEmail();
             var toName = returnRequest.Customer.GetFullName();
 
 			if (toEmail.IsEmpty())
 				return 0;
 
-            return SendNotification(messageTemplate, emailAccount, languageId, tokens, toEmail, toName);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         #endregion
@@ -1163,7 +1140,8 @@ namespace SmartStore.Services.Messages
             _messageTokenProvider.AddCustomerTokens(ctx.Tokens, ctx.Customer);
             _messageTokenProvider.AddStoreTokens(ctx.Tokens, _storeService.GetStoreById(ctx.StoreId.Value));
 
-            ctx.LanguageId = EnsureLanguageIsActive(ctx.LanguageId.Value, ctx.StoreId.Value);
+			var language = EnsureLanguageIsActive(ctx.LanguageId.Value, ctx.StoreId.Value);
+			ctx.LanguageId = language.Id;
 
             var messageTemplate = GetActiveMessageTemplate(messageTemplateName, ctx.StoreId.Value);
             if (messageTemplate == null)
@@ -1204,7 +1182,7 @@ namespace SmartStore.Services.Messages
 			if (store == null)
 				store = _storeContext.CurrentStore;
 
-			languageId = EnsureLanguageIsActive(languageId, store.Id);
+			var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("GiftCard.Notification", store.Id);
 			if (messageTemplate == null)
@@ -1217,12 +1195,11 @@ namespace SmartStore.Services.Messages
             
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = giftCard.RecipientEmail;
             var toName = giftCard.RecipientName;
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -1231,14 +1208,13 @@ namespace SmartStore.Services.Messages
         /// <param name="productReview">Product review</param>
         /// <param name="languageId">Message language identifier</param>
         /// <returns>Queued email identifier</returns>
-        public virtual int SendProductReviewNotificationMessage(ProductReview productReview,
-            int languageId)
+        public virtual int SendProductReviewNotificationMessage(ProductReview productReview, int languageId)
         {
             if (productReview == null)
                 throw new ArgumentNullException("productReview");
 
 			var store = _storeContext.CurrentStore;
-			languageId = EnsureLanguageIsActive(languageId, store.Id);
+			var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("Product.ProductReview", store.Id);
 			if (messageTemplate == null)
@@ -1252,17 +1228,14 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
 
             // use customer email as reply address
 			var replyTo = GetReplyToEmail(productReview.Customer);
 
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName,
-				replyTo.Item1, replyTo.Item2);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName, replyTo.Item1, replyTo.Item2);
         }
 
         /// <summary>
@@ -1277,7 +1250,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("product");
 
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("QuantityBelow.StoreOwnerNotification", store.Id);
 			if (messageTemplate == null)
@@ -1285,17 +1258,16 @@ namespace SmartStore.Services.Messages
 
 			var tokens = new List<Token>();
 			_messageTokenProvider.AddStoreTokens(tokens, store);
-			_messageTokenProvider.AddProductTokens(tokens, product, languageId);
+			_messageTokenProvider.AddProductTokens(tokens, product, language);
 
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         /// <summary>
@@ -1306,14 +1278,13 @@ namespace SmartStore.Services.Messages
         /// <param name="vatAddress">Received VAT address</param>
         /// <param name="languageId">Message language identifier</param>
         /// <returns>Queued email identifier</returns>
-        public virtual int SendNewVatSubmittedStoreOwnerNotification(Customer customer,
-            string vatName, string vatAddress, int languageId)
+        public virtual int SendNewVatSubmittedStoreOwnerNotification(Customer customer, string vatName, string vatAddress, int languageId)
         {
             if (customer == null)
                 throw new ArgumentNullException("customer");
 
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("NewVATSubmitted.StoreOwnerNotification", store.Id);
 			if (messageTemplate == null)
@@ -1329,17 +1300,14 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
 
             // use customer email as reply address
 			var replyTo = GetReplyToEmail(customer);
 
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName,
-				replyTo.Item1, replyTo.Item2);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName, replyTo.Item1, replyTo.Item2);
         }
 
         /// <summary>
@@ -1354,7 +1322,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("blogComment");
 
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("Blog.BlogComment", store.Id);
 			if (messageTemplate == null)
@@ -1368,17 +1336,14 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
 
             // use customer email as reply address
 			var replyTo = GetReplyToEmail(blogComment.Customer);
 
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName,
-				replyTo.Item1, replyTo.Item2);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName, replyTo.Item1, replyTo.Item2);
         }
 
         /// <summary>
@@ -1393,7 +1358,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("newsComment");
 
 			var store = _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 			
 			var messageTemplate = GetActiveMessageTemplate("News.NewsComment", store.Id);
 			if (messageTemplate == null)
@@ -1407,17 +1372,14 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var toEmail = emailAccount.Email;
             var toName = emailAccount.DisplayName;
 
             // use customer email as sender/reply address
 			var replyTo = GetReplyToEmail(newsComment.Customer);
 
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName,
-				replyTo.Item1, replyTo.Item2);
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName, replyTo.Item1, replyTo.Item2);
         }
 
         /// <summary>
@@ -1432,7 +1394,7 @@ namespace SmartStore.Services.Messages
                 throw new ArgumentNullException("subscription");
 
 			var store = _storeService.GetStoreById(subscription.StoreId) ?? _storeContext.CurrentStore;
-            languageId = EnsureLanguageIsActive(languageId, store.Id);
+            var language = EnsureLanguageIsActive(languageId, store.Id);
 
 			var messageTemplate = GetActiveMessageTemplate("Customer.BackInStock", store.Id);
 			if (messageTemplate == null)
@@ -1447,13 +1409,12 @@ namespace SmartStore.Services.Messages
             //event notification
             _eventPublisher.MessageTokensAdded(messageTemplate, tokens);
 
-            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, languageId);
+            var emailAccount = GetEmailAccountOfMessageTemplate(messageTemplate, language.Id);
             var customer = subscription.Customer;
             var toEmail = customer.Email;
             var toName = customer.GetFullName();
-            return SendNotification(messageTemplate, emailAccount,
-                languageId, tokens,
-                toEmail, toName);
+
+            return SendNotification(messageTemplate, emailAccount, language.Id, tokens, toEmail, toName);
         }
 
         #endregion